Switch from long long to int64_t
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@8560 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/turbine/turbine.h
@@ -42,7 +42,7 @@ typedef enum
   TURBINE_ACTION_WORK = 3
 } turbine_action_type;
 
-typedef long long turbine_transform_id;
+typedef int64_t turbine_transform_id;
 
 typedef struct {
   turbine_datum_id td;
